Refactor drag calculations to use raw GPS data instead of smoothed data
This commit updates the drag racing calculation logic to use raw GPS data directly, removing the dependency on the post-processing and smoothed data layers. This simplifies the calculation process and removes a previously complex and partially implemented step.

**Key Changes:**

*   **`DragTimeCalculation.kt`:**
    *   Calculations for `timeFromZeroToHundred` and `quarterMile` now fetch data from `rawGPSDataDao` instead of `smoothedDataDao`.
    *   The post-processing step (`postProcessing()` function) has been commented out and is no longer called, as calculations are now performed on raw data.
    *   The `speed` property from `RawGPSData` is used for calculations instead of `smoothedSpeed`.
*   **`RawGPSDataDao.kt`:**
    *   The `getGPSDataBySession` function's `sessionId` parameter was changed from `Int` to `Long?` to align with the data type of the session ID.
*   **`GraphScreen.kt` & `CanvasMap.kt`:**
    *   Updated the graph screen and map drawing functions to consume `RawGPSData` instead of `SmoothedGPSData` for displaying session paths and speed graphs.
*   **`DragScreen.kt`:**
    *   The call to `endSessionPostProcess` has been removed as it is no longer needed.
*   **`PostProcessing.kt`:**
    *   The method to retrieve GPS data was updated to use the new `Long?` session ID type.

+++ app/src/main/java/com/example/trackpro/CalculationClasses/PostProcessing.kt
@@ -1,15 +1,11 @@
 package com.example.trackpro.CalculationClasses
 
-import android.nfc.Tag
 import android.util.Log
 import com.example.trackpro.DataClasses.RawGPSData
 import com.example.trackpro.DataClasses.SmoothedGPSData
 import com.example.trackpro.DataClasses.TrackCoordinatesData
 import com.example.trackpro.ESPDatabase
-import kotlinx.coroutines.CoroutineScope
-import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.first
-import kotlinx.coroutines.launch
 import kotlin.math.atan2
 import kotlin.math.cos
 import kotlin.math.sin
@@ -19,7 +15,7 @@ class PostProcessing(val database: ESPDatabase) {
 
     suspend fun postProcessing(sessionId: Long) {
         // Step 1: Retrieve raw GPS data for the session
-        val sessionItems: List<RawGPSData> = database.rawGPSDataDao().getGPSDataBySession(sessionId.toInt())
+        val sessionItems: List<RawGPSData> = database.rawGPSDataDao().getGPSDataBySession(sessionId)
 
         // Step 2: Apply smoothing
         val windowSize = 15
